pyproject.toml
[project]
name = "nkazureblobman"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.28.0",
    "dotenv>=0.9.9",
    "msal>=1.34.0",
]
